Separate assertions and decorators into separate modules
