+++ docs/pages/api_documentation/private/assertions.rst
@@ -1,9 +1,8 @@
-Decorators
+Assertions
 ____________________________________________________
-.. automodule:: gisnav._decorators
+.. automodule:: gisnav._assertions
    :autosummary:
    :members:
    :undoc-members:
    :special-members: __init__
    :show-inheritance:
-   :exclude-members: P
